Move testing code for generating credentials somewhere more public.
