+++ flocker/ca/functional/test_validation.py
@@ -6,31 +6,22 @@ Test validation of keys generated by flocker-ca.
 
 from __future__ import print_function
 
-from tempfile import mkdtemp
-from shutil import rmtree
-import atexit
-
-from pyrsistent import PRecord, field
-
 from OpenSSL.SSL import Context, TLSv1_METHOD, Error as SSLError
 
 from twisted.trial.unittest import TestCase
-from twisted.internet.ssl import PrivateCertificate
 from twisted.internet.endpoints import (
     SSL4ServerEndpoint, connectProtocol, SSL4ClientEndpoint,
     )
 from twisted.internet import reactor
 from twisted.internet.defer import Deferred
-from twisted.python.filepath import FilePath
 from twisted.internet.protocol import Protocol, ServerFactory
 
 
 from ...testtools import find_free_port
-from .._ca import (
-    RootCredential, UserCredential, NodeCredential, ControlCredential)
 from .._validation import (
     ControlServicePolicy, amp_server_context_factory, rest_api_context_factory,
     )
+from ..testtools import get_credential_sets
 
 
 EXPECTED_STRING = b"Mr. Watson, come here; I want to see you."
@@ -66,50 +57,6 @@ class ReceivingProtocol(Protocol):
         if self.result is not None:
             self.result.errback(reason)
             self.result = None
-
-
-class CredentialSet(PRecord):
-    """
-    A full set of credentials for a CA.
-
-    :ivar RootCredential root: The CA root credential.
-    :ivar ControlCredential control: A control service credential.
-    :ivar UserCredential user: A user credential.
-    :ivar NodeCredential node: A CA root credentials.
-    """
-    root = field()
-    control = field()
-    user = field()
-    node = field()
-
-    @staticmethod
-    def create():
-        """
-        :return: A new ``CredentialSet``.
-        """
-        directory = FilePath(mkdtemp())
-        atexit.register(rmtree, directory.path)
-        root = RootCredential.initialize(directory, b"mycluster")
-        user = UserCredential.initialize(directory, root, u"allison")
-        node = NodeCredential.initialize(directory, root)
-        control = ControlCredential.initialize(directory, root, b"127.0.0.1")
-        return CredentialSet(root=root, user=user, node=node, control=control)
-
-
-# Cached CredentialSets generated by get_sets():
-_credential_set1 = None
-_credential_set2 = None
-
-
-def get_sets():
-    """
-    :return: Tuple, a pair of ``CredentialSet`` instances.
-    """
-    global _credential_set1, _credential_set2
-    if _credential_set1 is None:
-        _credential_set1 = CredentialSet.create()
-        _credential_set2 = CredentialSet.create()
-    return _credential_set1, _credential_set2
 
 
 class PeerContextFactory(object):
@@ -160,7 +107,7 @@ def make_validation_tests(context_factory_fixture,
         :ivar CertificateSet another_ca: A different CA's certificates.
         """
         def setUp(self):
-            self.good_ca, self.another_ca = get_sets()
+            self.good_ca, self.another_ca = get_credential_sets()
 
         def _handshake(self, credential):
             """
