+++ flocker/ca/testtools.py
@@ -4,7 +4,19 @@
 Testing utilities for ``flocker.ca``.
 """
 
+import atexit
+from tempfile import mkdtemp
+from shutil import rmtree
+
 from OpenSSL.crypto import X509Extension
+
+from twisted.python.filepath import FilePath
+
+from pyrsistent import PRecord, field
+
+from ..ca import (
+    RootCredential, ControlCredential, NodeCredential, UserCredential,
+)
 
 
 def assert_has_extension(test, credential, name, value):
@@ -29,3 +41,51 @@ def assert_has_extension(test, credential, name, value):
             test.assertEqual(extension.get_data(), expected.get_data())
             return
     test.fail("Couldn't find extension {}.".format(name))
+
+
+class CredentialSet(PRecord):
+    """
+    A full set of credentials for a CA.
+
+    :ivar RootCredential root: The CA root credential.
+    :ivar ControlCredential control: A control service credential.
+    :ivar UserCredential user: A user credential.
+    :ivar NodeCredential node: A CA root credentials.
+    """
+    root = field()
+    control = field()
+    user = field()
+    node = field()
+
+    @staticmethod
+    def create():
+        """
+        :return: A new ``CredentialSet``.
+        """
+        directory = FilePath(mkdtemp())
+        atexit.register(rmtree, directory.path)
+        root = RootCredential.initialize(directory, b"mycluster")
+        user = UserCredential.initialize(directory, root, u"allison")
+        node = NodeCredential.initialize(directory, root)
+        control = ControlCredential.initialize(directory, root, b"127.0.0.1")
+        return CredentialSet(root=root, user=user, node=node, control=control)
+
+
+# Cached CredentialSets generated by get_sets():
+_credential_set1 = None
+_credential_set2 = None
+
+
+def get_credential_sets():
+    """
+    Get a pair of cached sets of CA-specific credentials.
+
+    Since they are cached across tests they should not be modified.
+
+    :return: Tuple, a pair of ``CredentialSet`` instances.
+    """
+    global _credential_set1, _credential_set2
+    if _credential_set1 is None:
+        _credential_set1 = CredentialSet.create()
+        _credential_set2 = CredentialSet.create()
+    return _credential_set1, _credential_set2
